+++ tibiapy/parsers/world.py
@@ -1,12 +1,13 @@
+from __future__ import annotations
 import re
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Optional, List
 
 from tibiapy.builders.world import (WorldBuilder, WorldEntryBuilder,
                                     WorldOverviewBuilder)
 from tibiapy.enums import (BattlEyeType, PvpType,
                            TransferType, WorldLocation)
 from tibiapy.errors import InvalidContent
-from tibiapy.models import OnlineCharacter, World
+from tibiapy.models import OnlineCharacter, World, WorldOverview, WorldEntry
 from tibiapy.utils import (parse_integer, parse_tibia_datetime,
                            parse_tibia_full_date, parse_tibiacom_content,
                            try_enum, parse_tables_map, get_rows, clean_text)
@@ -24,9 +25,10 @@ battleye_regexp = re.compile(r'since ([^.]+).')
 
 
 class WorldParser:
+    """Parses Tibia.com content into worlds."""
 
     @classmethod
-    def from_content(cls, content):
+    def from_content(cls, content: str) -> Optional[World]:
         """Parse a Tibia.com response into a :class:`World`.
 
         Parameters
@@ -65,52 +67,51 @@ class WorldParser:
         return builder.build()
 
     @classmethod
-    def _parse_world_info(cls, builder: WorldBuilder, world_info_table):
+    def _parse_world_info(cls, builder: WorldBuilder, world_info_table: bs4.Tag):
         """
         Parse the World Information table from Tibia.com and adds the found values to the object.
 
         Parameters
         ----------
         world_info_table: :class:`bs4.Tag`
+            The table containing the world's information.
         """
+        field_actions = {
+            "Status": lambda v: builder.is_online("online" in v.lower()),
+            "Players Online": lambda v: builder.online_count(parse_integer(v)),
+            "Online Record": lambda v: cls._parse_online_record(builder, v),
+            "Creation Date": lambda v: cls._parse_creation_date(builder, v),
+            "Location": lambda v: builder.location(try_enum(WorldLocation, v)),
+            "PvP Type": lambda v: builder.pvp_type(try_enum(PvpType, v)),
+            "Premium Type": lambda v: builder.is_premium_only(True),
+            "Transfer Type": lambda v: builder.transfer_type(try_enum(TransferType, v, TransferType.REGULAR)),
+            "World Quest Titles": lambda v: cls._parse_world_quest_titles(builder, v),
+            "BattlEye Status": lambda v: cls._parse_battleye_status(builder, v),
+            "Game World Type": lambda v: builder.is_experimental(v.lower() == "experimental"),
+        }
         for row in get_rows(world_info_table):
             cols_raw = row.select('td')
             cols = [clean_text(ele) for ele in cols_raw]
             field, value = cols
             field = field.replace(":", "")
-            if field == "Status":
-                builder.is_online("online" in value.lower())
-            elif field == "Players Online":
-                builder.online_count(parse_integer(value))
-            elif field == "Online Record":
-                cls._parse_online_record(builder, value)
-            elif field == "Creation Date":
-                cls._parse_creation_date(builder, value)
-            elif field == "Location":
-                builder.location(try_enum(WorldLocation, value))
-            elif field == "PvP Type":
-                builder.pvp_type(try_enum(PvpType, value))
-            elif field == "Premium Type":
-                builder.is_premium_only(True)
-            elif field == "Transfer Type":
-                builder.transfer_type(try_enum(TransferType, value, TransferType.REGULAR))
-            elif field == "World Quest Titles":
-                titles = [q.strip() for q in value.split(",")]
-                if "currently has no title" not in titles[0]:
-                    builder.world_quest_titles(titles)
-            elif field == "BattlEye Status":
-                cls._parse_battleye_status(builder, value)
-            elif field == "Game World Type":
-                builder.experimental(value.lower() == "experimental")
+            if field in field_actions:
+                action = field_actions[field]
+                action(value)
 
     @classmethod
-    def _parse_online_record(cls, builder: WorldBuilder, value):
+    def _parse_world_quest_titles(cls, builder: WorldBuilder, value: str):
+        titles = [q.strip() for q in value.split(",")]
+        if "currently has no title" not in titles[0]:
+            builder.world_quest_titles(titles)
+
+    @classmethod
+    def _parse_online_record(cls, builder: WorldBuilder, value: str):
         if m := record_regexp.match(value):
             builder.record_count(parse_integer(m.group("count")))
             builder.record_date(parse_tibia_datetime(m.group("date")))
 
     @classmethod
-    def _parse_creation_date(cls, builder: WorldBuilder, value):
+    def _parse_creation_date(cls, builder: WorldBuilder, value: str):
         month, year = value.split("/")
         month = int(month)
         year = int(year)
@@ -118,26 +119,28 @@ class WorldParser:
         builder.creation_date(f"{year:d}-{month:02d}")
 
     @classmethod
-    def _parse_battleye_status(cls, builder, battleye_string):
+    def _parse_battleye_status(cls, builder: WorldBuilder, battleye_string: str):
         """Parse the BattlEye string and applies the results.
 
         Parameters
         ----------
+        builder: :class:`WorldBuilder`
+            The builder instance used to set the values.
         battleye_string: :class:`str`
-            String containing the world's Battleye Status.
+            String containing the world's BattlEye Status.
         """
         if m := battleye_regexp.search(battleye_string):
             battleye_date = parse_tibia_full_date(m.group(1))
-            builder.battleye_since(battleye_date)\
-                .battleye_type(BattlEyeType.PROTECTED if battleye_date else BattlEyeType.INITIALLY_PROTECTED)
+            (builder.battleye_since(battleye_date)
+             .battleye_type(BattlEyeType.PROTECTED if battleye_date else BattlEyeType.INITIALLY_PROTECTED)
+             )
         else:
-            builder.battleye_since(None)\
-                .battleye_type(BattlEyeType.UNPROTECTED)
+            builder.battleye_since(None).battleye_type(BattlEyeType.UNPROTECTED)
 
 
 class WorldOverviewParser:
     @classmethod
-    def from_content(cls, content):
+    def from_content(cls, content: str) -> WorldOverview:
         """Parse the content of the World Overview section from Tibia.com into an object of this class.
 
         Parameters
@@ -157,8 +160,7 @@ class WorldOverviewParser:
         """
         parsed_content = parse_tibiacom_content(content)
         try:
-            record_table, *tables \
-                = parsed_content.select("table.TableContent")
+            record_table, *tables = parsed_content.select("table.TableContent")
             m = record_regexp.search(record_table.text)
             return (WorldOverviewBuilder()
                     .record_count(parse_integer(m.group("count")))
@@ -169,7 +171,7 @@ class WorldOverviewParser:
             raise InvalidContent("content does not belong to the World Overview section in Tibia.com", e) from e
 
     @classmethod
-    def _parse_worlds(cls, world_rows):
+    def _parse_worlds(cls, world_rows: List[bs4.Tag]) -> List[WorldEntry]:
         """Parse the world columns and adds the results to :py:attr:`worlds`.
 
         Parameters
@@ -181,10 +183,10 @@ class WorldOverviewParser:
         for world_row in world_rows:
             cols = world_row.select("td")
             name = cols[0].text.strip()
-            online = True
+            is_online = True
             online_count = parse_integer(cols[1].text.strip(), None)
             if online_count is None:
-                online = False
+                is_online = False
                 online_count = 0
             location = try_enum(WorldLocation, cols[2].text.replace("\u00a0", " ").strip())
             pvp = try_enum(PvpType, cols[3].text.strip())
@@ -193,7 +195,7 @@ class WorldOverviewParser:
                        .location(location)
                        .pvp_type(pvp)
                        .online_count(online_count)
-                       .is_online(online))
+                       .is_online(is_online))
             # Check Battleye icon to get information
             battleye_icon = cols[4].select_one("span.HelperDivIndicator")
             if battleye_icon is not None:
@@ -206,18 +208,18 @@ class WorldOverviewParser:
         return worlds
 
     @classmethod
-    def _parse_additional_info(cls, builder, additional_info):
+    def _parse_additional_info(cls, builder: WorldEntryBuilder, additional_info: str):
         if "blocked" in additional_info:
             builder.transfer_type(TransferType.BLOCKED)
         elif "locked" in additional_info:
             builder.transfer_type(TransferType.LOCKED)
         else:
             builder.transfer_type(TransferType.REGULAR)
-        builder.experimental("experimental" in additional_info)
+        builder.is_experimental("experimental" in additional_info)
         builder.is_premium_only("premium" in additional_info)
 
     @classmethod
-    def _parse_worlds_tables(cls, tables):
+    def _parse_worlds_tables(cls, tables: List[bs4.Tag]):
         """Parse the tables and adds the results to the world list.
 
         Parameters